requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
faiss-cpu
python-dotenv
sentence-transformers
transformers
huggingface-hub
safetensors